mock scan and toast
